+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.0-beta2] - 2023-10-31
+- Fixed: command registration error
+
 ## [2.0.0-beta] - 2023-10-31
 - Added: support for contao 5
 - Added: support for utils bundle v3